Implement both tasks
ESR - add sysconfig to ignore non-credit lines
https://github.com/metasfresh/metasfresh/issues/2118
ESR - Show actual number of transactions
https://github.com/metasfresh/metasfresh/issues/2119
